+++ main.tf
@@ -2,15 +2,17 @@ resource "azurerm_application_gateway" "appgw" {
   name                = var.name
   resource_group_name = var.resource_group_name
   location            = var.location
+  tags                = var.tags
 
   sku {
     name     = var.sku.size
     tier     = var.sku.tier
-    capacity = lookup(var.sku, "capacity", null)
+    capacity = var.sku.capacity
   }
 
   dynamic "autoscale_configuration" {
-    for_each = var.autoscale_configuration != {} ? [""] : []
+    for_each = var.autoscale_configuration != null ? [""] : []
+
     content {
       min_capacity = var.autoscale_configuration.min_capacity
       max_capacity = var.autoscale_configuration.max_capacity
@@ -24,15 +26,17 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "waf_configuration" {
     for_each = local.waf_configuration_enabled ? [""] : []
+
     content {
       enabled          = var.waf_configuration.enabled
-      firewall_mode    = lookup(var.waf_configuration, "firewall_mode", "Detection")
-      rule_set_version = lookup(var.waf_configuration, "rule_set_version", "3.0")
+      firewall_mode    = var.waf_configuration.firewall_mode
+      rule_set_version = var.waf_configuration.rule_set_version
     }
   }
 
   dynamic "frontend_ip_configuration" {
     for_each = local.public_ip_address_id ? [""] : []
+
     content {
       name                 = "Public-frontend-ip-configuration"
       public_ip_address_id = var.frontend_ip_configuration.public_ip_address_id
@@ -41,6 +45,7 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "frontend_ip_configuration" {
     for_each = local.private_ip_address || local.private_ip_address_allocation ? [""] : []
+
     content {
       name                          = "Private-frontend-ip-configuration"
       subnet_id                     = var.subnet_id
@@ -51,9 +56,10 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "backend_address_pool" {
     for_each = var.backend_address_pools
+
     content {
       name         = backend_address_pool.value.name
-      ip_addresses = lookup(backend_address_pool.value, "ip_addresses", "") == "" ? null : split(",", backend_address_pool.value.ip_addresses)
+      ip_addresses = backend_address_pool.value.ip_addresses
     }
   }
 
@@ -69,24 +75,27 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "identity" {
     for_each = var.identity_id != null ? [""] : []
+
     content {
       type         = "UserAssigned"
       identity_ids = [var.identity_id]
     }
   }
 
-  dynamic "ssl_certificate" {
-    for_each = var.ssl_certificates
-    content {
-      name                = ssl_certificate.value.name
-      data                = lookup(ssl_certificate.value, "data", null)
-      password            = lookup(ssl_certificate.value, "password", null)
-      key_vault_secret_id = lookup(ssl_certificate.value, "key_vault_secret_id", null)
-    }
-  }
+  # dynamic "ssl_certificate" {
+  #   for_each = var.ssl_certificates
+
+  #   content {
+  #     name                = ssl_certificate.value.name
+  #     data                = ssl_certificate.value.data
+  #     password            = ssl_certificate.value.password
+  #     key_vault_secret_id = ssl_certificate.value.key_vault_secret_id
+  #   }
+  # }
 
   dynamic "http_listener" {
     for_each = var.http_listeners
+
     content {
       name                           = http_listener.value.name
       frontend_ip_configuration_name = "${http_listener.value.frontend_ip_configuration}-frontend-ip-configuration"
@@ -99,9 +108,10 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "probe" {
     for_each = var.probes
+
     content {
       name                = probe.value.name
-      host                = lookup(probe.value, "host", null)
+      host                = probe.value.host
       protocol            = probe.value.protocol
       path                = probe.value.path
       interval            = probe.value.interval
@@ -112,19 +122,21 @@ resource "azurerm_application_gateway" "appgw" {
 
   dynamic "backend_http_settings" {
     for_each = var.backend_http_settings
+
     content {
       cookie_based_affinity = "Disabled"
       name                  = backend_http_settings.value.name
       port                  = backend_http_settings.value.port
       protocol              = backend_http_settings.value.protocol
       request_timeout       = backend_http_settings.value.request_timeout
-      host_name             = lookup(backend_http_settings.value, "host_name", null)
-      probe_name            = lookup(backend_http_settings.value, "probe_name", null)
+      host_name             = backend_http_settings.value.host_name
+      probe_name            = backend_http_settings.value.probe_name
     }
   }
 
   dynamic "request_routing_rule" {
     for_each = var.request_routing_rules
+
     content {
       name                       = request_routing_rule.value.name
       rule_type                  = "Basic"
@@ -133,6 +145,4 @@ resource "azurerm_application_gateway" "appgw" {
       backend_http_settings_name = request_routing_rule.value.backend_http_settings_name
     }
   }
-
-  tags = var.tags
 }